requirements.txt
arrow==0.10.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-appconf==1.0.2
django-crispy-forms==1.6.1
idna==2.6
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
